+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "lib/nanoprintf"]
-	path = lib/nanoprintf
+	path = include/nanoprintf
 	url = https://github.com/charlesnicholson/nanoprintf
