+++ Makefile
@@ -7,7 +7,8 @@ TARGET 		= aarch64-unknown-none-elf
 CFLAGS 		= -Wall -Wextra -Wshadow \
 			  -ffreestanding \
 			  -mcpu=cortex-a53 --target=$(TARGET) \
-			  -nostdlib -Os -Ilib -D__CLANG_INTTYPES_H
+			  -nostdlib -Os \
+			  -Iinclude
 QEMU_FLAGS 	= -display none \
 			  -serial null -serial stdio \
 			  -smp cpus=4
