+++ compile_flags.txt
@@ -1,1 +1,1 @@
--Ilib
+-Iinclude
