+++ include/cmd.h
